Generate the Toolbox layout ABI
(cherry picked from commit 3a031b3c3e3324cd81c096b903a590fd4871f023)
