+++ src/renderer/toolbox.js
@@ -1,43 +1,10 @@
-const MAGIC = 0x42545747;
-const SNAPSHOT_ABI = 1;
-const SNAPSHOT_BYTES = 64;
-const CONFIG_FIELDS = Object.freeze([
-  "contextRoot",
-  "agentArray",
-  "targetAgentId",
-  "gameContextSlot",
-  "characterContext",
-  "mapId",
-  "isExplorable",
-  "currentMapId",
-  "currentInstanceType",
-  "playerNumber",
-  "agentId",
-  "agentX",
-  "agentY",
-  "agentType",
-  "agentPlayerNumber",
-  "agentModelType",
-]);
-const INSTANCE_NAMES = Object.freeze(["Outpost", "Explorable", "Loading"]);
-const RANGE_NAMES = Object.freeze([
-  "None",
-  "Adjacent",
-  "Nearby",
-  "Area",
-  "Earshot",
-  "Spellcast",
-  "Spirit",
-  "Compass",
-  "Beyond compass",
-]);
+import {
+  readToolboxSnapshot,
+  TOOLBOX_SNAPSHOT_ABI,
+  TOOLBOX_SNAPSHOT_BYTES,
+} from "./toolbox-snapshot.js";
 
-const FLAGS = Object.freeze({
-  ready: 1 << 0,
-  player: 1 << 1,
-  target: 1 << 2,
-  loading: 1 << 3,
-});
+const CONFIG_WORDS = 16;
 
 function decodeManifest(module) {
   const sections = WebAssembly.Module.customSections(module, "toolbox_manifest");
@@ -45,13 +12,16 @@ function decodeManifest(module) {
   try {
     const value = JSON.parse(new TextDecoder().decode(sections[0]));
     if (
-      value?.transformAbi !== 1
-      || value?.snapshotAbi !== SNAPSHOT_ABI
-      || value?.snapshotBytes !== SNAPSHOT_BYTES
-      || value?.configBytes !== CONFIG_FIELDS.length * 4
+      value?.transformAbi !== 2
+      || value?.snapshotAbi !== TOOLBOX_SNAPSHOT_ABI
+      || value?.snapshotBytes !== TOOLBOX_SNAPSHOT_BYTES
+      || value?.configBytes !== CONFIG_WORDS * 4
       || !Number.isInteger(value?.tableSlot)
-      || !value?.layout
-      || CONFIG_FIELDS.some((field) => !Number.isInteger(value.layout[field]))
+      || !Array.isArray(value?.layoutWords)
+      || value.layoutWords.length !== CONFIG_WORDS
+      || value.layoutWords.some(
+        (word) => !Number.isInteger(word) || word < 0 || word > 0xffff_ffff,
+      )
     ) {
       return null;
     }
@@ -59,87 +29,6 @@ function decodeManifest(module) {
   } catch {
     return null;
   }
-}
-
-function stableFloat(value) {
-  return Number.isFinite(value) ? value : 0;
-}
-
-export function readToolboxSnapshot(buffer, pointer) {
-  if (
-    !(buffer instanceof ArrayBuffer)
-    || !Number.isInteger(pointer)
-    || pointer < 0
-    || pointer + SNAPSHOT_BYTES > buffer.byteLength
-  ) {
-    return Object.freeze({ status: "waiting", reason: "memory" });
-  }
-  const view = new DataView(buffer, pointer, SNAPSHOT_BYTES);
-  const firstSequence = view.getUint32(8, true);
-  if ((firstSequence & 1) !== 0) {
-    return Object.freeze({ status: "waiting", reason: "writing" });
-  }
-  const magic = view.getUint32(0, true);
-  const abi = view.getUint16(4, true);
-  const byteLength = view.getUint16(6, true);
-  const flags = view.getUint32(12, true);
-  const state = {
-    sequence: firstSequence,
-    tickCount: view.getUint32(16, true),
-    mapId: view.getUint32(20, true),
-    instanceType: view.getUint32(24, true),
-    playerId: view.getUint32(28, true),
-    playerX: stableFloat(view.getFloat32(32, true)),
-    playerY: stableFloat(view.getFloat32(36, true)),
-    targetId: view.getUint32(40, true),
-    targetType: view.getUint32(44, true),
-    targetX: stableFloat(view.getFloat32(48, true)),
-    targetY: stableFloat(view.getFloat32(52, true)),
-    distance: stableFloat(view.getFloat32(56, true)),
-    rangeBand: view.getUint32(60, true),
-  };
-  const secondSequence = view.getUint32(8, true);
-  if (
-    magic !== MAGIC
-    || abi !== SNAPSHOT_ABI
-    || byteLength !== SNAPSHOT_BYTES
-    || firstSequence !== secondSequence
-    || (secondSequence & 1) !== 0
-  ) {
-    return Object.freeze({ status: "waiting", reason: "snapshot" });
-  }
-  if ((flags & FLAGS.loading) !== 0) {
-    return Object.freeze({
-      status: "waiting",
-      reason: "loading",
-      sequence: secondSequence,
-      tickCount: state.tickCount,
-    });
-  }
-  if ((flags & (FLAGS.ready | FLAGS.player)) !== (FLAGS.ready | FLAGS.player)) {
-    return Object.freeze({
-      status: "waiting",
-      reason: "game",
-      sequence: secondSequence,
-      tickCount: state.tickCount,
-    });
-  }
-  const targetValid = (flags & FLAGS.target) !== 0;
-  return Object.freeze({
-    status: "ready",
-    ...state,
-    instanceName: INSTANCE_NAMES[state.instanceType] ?? "Unknown",
-    targetValid,
-    targetKind: targetValid ? agentKind(state.targetType) : "None",
-    rangeName: RANGE_NAMES[state.rangeBand] ?? "Unknown",
-  });
-}
-
-function agentKind(type) {
-  if ((type & 0x400) !== 0) return "Item";
-  if ((type & 0x200) !== 0) return "Gadget";
-  if ((type & 0xdb) !== 0) return "Living";
-  return "Unknown";
 }
 
 function text(id, value) {
@@ -286,17 +175,15 @@ async function install(instance, module) {
   let configPointer = 0;
   let stopRenderer = () => {};
   try {
-    snapshotPointer = Number(exports.malloc(SNAPSHOT_BYTES));
-    configPointer = Number(exports.malloc(CONFIG_FIELDS.length * 4));
+    snapshotPointer = Number(exports.malloc(TOOLBOX_SNAPSHOT_BYTES));
+    configPointer = Number(exports.malloc(manifest.configBytes));
     if (!snapshotPointer || !configPointer) throw new Error("Toolbox allocation failed");
     const config = new Uint32Array(
       exports.memory.buffer,
       configPointer,
-      CONFIG_FIELDS.length,
+      manifest.layoutWords.length,
     );
-    CONFIG_FIELDS.forEach((field, index) => {
-      config[index] = manifest.layout[field];
-    });
+    config.set(manifest.layoutWords);
 
     const kernelBytes = await (await fetch("toolbox-kernel.wasm")).arrayBuffer();
     const kernel = await WebAssembly.instantiate(kernelBytes, {
@@ -308,9 +195,9 @@ async function install(instance, module) {
       || typeof kernel.instance.exports.toolbox_tick !== "function"
       || kernel.instance.exports.toolbox_init(
         snapshotPointer,
-        SNAPSHOT_BYTES,
+        TOOLBOX_SNAPSHOT_BYTES,
         configPointer,
-        CONFIG_FIELDS.length * 4,
+        manifest.configBytes,
       ) !== 1
     ) {
       throw new Error("Toolbox kernel rejected its ABI");
